Added an attempt to build ADCIRC as a step.
Issue #127: Also added a "setup-env" step and partitined out
the setting of specific environmental variables to the appropriate
step. Added a MACHINENAME variable that reflects the value of
the required --machinename flag..

---

Added step to install required Python libraries.

Issue #136: Covers pika, netCDF5, and the pptx modules. Did not verify
that RabbitMQ support is provided.