Deprecating Date::Pcalc in favor of Date::Calc.
Issue #140: Removed installation of Date::Pcalc because it required
being treated differently than the other modules and its tests were
failing. Replaced all instances of Date::Pcalc in scripts with
Date::Calc. This commit does NOT remove the old version(s) of Date::Pcalc
that are distributed with ASGS, but that (and removing all CPAN
modules that are distributed) is the next step since the installation
of all non-custom modules is being handled using perlbrew and cpanm.