Attempt to get derived information based on INSTANCENAME in config
Using INSTANCENAME that can be grep'd out of the config file, we
can discover the location of the state file in the scratch dir; from
there we can discover the RUNDIR, SYSLOG, and location of the primary
run.properties file. Using this information, implemented 'watchlog'
that 'tail -f's the log file. No derived information is save in the
named profile since it should be discovered by first inspecting the
configuration file.