Adding foundational support for asgsctl.
WIP

Issue #378: Added support to asgs_main.sh to add a process id file
(to $RUNDIR/asgs_main.pid). Added initial tool (asgsctl) to create
and manage touch files that asgs_main.sh will then be able to use
to do the needful when detecting touchfiles in $RUNDIR.

Support in asgs_main.sh has not been added and will be done so in a
under a different issue number.

Added $SCRIPTDIR/bin to PATH in asgsh. This is also meant for providing
a space for extending asgsh via external and contributed scripts.